Fixed data type mismatch causing dashboard bubbles to always be green

+++ src/modules/clients/controller.js
@@ -272,10 +272,10 @@ export const getClients = async (req, res) => {
         c.integrity_status, c.current_root_hash, c.last_reviewed_at,
         c.attestation_valid, c.file_count, c.last_heartbeat, c.last_boot_id,
         CASE WHEN COUNT(CASE WHEN e.reviewed = false AND e.event_type = 'attestation_failed' THEN 1 END) > 0 THEN 'FAILED' ELSE 'OK' END as attestation_status,
-        COUNT(CASE WHEN e.reviewed = false AND e.event_type NOT IN ('heartbeat', 'heartbeat_missed', 'directory_selected', 'directory_unselected', 'registration', 'deregistration', 'reinstall', 'uninstall', 'attestation_failed') THEN 1 END) as integrity_change_count,
-        COUNT(CASE WHEN e.reviewed = false AND e.event_type = 'heartbeat_missed' THEN 1 END) as missed_heartbeat_count,
-        COUNT(CASE WHEN e.reviewed = false AND e.event_type = 'attestation_failed' THEN 1 END) as attestation_error_count,
-        COUNT(CASE WHEN e.reviewed = false THEN 1 END) as unreviewed_events,
+        COUNT(CASE WHEN e.reviewed = false AND e.event_type NOT IN ('heartbeat', 'heartbeat_missed', 'directory_selected', 'directory_unselected', 'registration', 'deregistration', 'reinstall', 'uninstall', 'attestation_failed') THEN 1 END)::int as integrity_change_count,
+        COUNT(CASE WHEN e.reviewed = false AND e.event_type = 'heartbeat_missed' THEN 1 END)::int as missed_heartbeat_count,
+        COUNT(CASE WHEN e.reviewed = false AND e.event_type = 'attestation_failed' THEN 1 END)::int as attestation_error_count,
+        COUNT(CASE WHEN e.reviewed = false THEN 1 END)::int as unreviewed_events,
         MAX(e.timestamp) as last_event
       FROM clients c
       LEFT JOIN events e ON c.client_id = e.client_id
